A cast to float must of accidently got deleted in a previous update. This was causing some seemingly benign array out of bounds exceptions to be output to the console. Fixed.
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@9258 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/src/main/java/weka/gui/AttributeVisualizationPanel.java
@@ -765,7 +765,7 @@ public class AttributeVisualizationPanel
             
             try {
               //1. see footnote at the end of this file
-              t =(int) Math.ceil((
+              t =(int) Math.ceil((float)(
               (m_data.instance(k).value(m_attribIndex)-m_as.numericStats.min)
               / barRange));
               if(t==0) {
